+++ manual/Website1/src/App.jsx
@@ -1,7 +1,6 @@
 import React, { Component } from 'react';
-import { ExternalComponent } from '../../../index';
+import { ExternalComponent,corsImport } from 'webpack-external-import';
 import HelloWorld from './components/goodbye-world';
-import corsImport from "../../../src/corsImport";
 
 import('moment');
 class App extends Component {
@@ -11,7 +10,7 @@ class App extends Component {
   }
 
   componentDidMount() {
-    corsImport(/* webpackIgnore:true */'http://localhost:3002/importManifest.js').then(() => {
+    corsImport('http://localhost:3002/importManifest.js').then(() => {
       this.setState({ manifestLoaded: true });
       import(/* webpackIgnore:true */`http://localhost:3002/${window.entryManifest['website-two']['SomeExternalModule.js']}`).then(() => {
         console.log('got module, will render it in 2 seconds');
