+++ manual/Website1/src/index.jsx
@@ -1,3 +1,4 @@
+import 'webpack-external-import/polyfill';
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App.jsx';
