+++ src/polyfill.js
@@ -1,0 +1,5 @@
+import dimport from 'dimport/legacy';
+
+if (!window.import) {
+  window.import = dimport;
+}
